refactor: rewrite namespaces
