+++ CleannetCode_bot/Features/Homework/models/GithubData.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace CleannetCode_bot.Features.Homework;
+namespace CleannetCode_bot.Features.Homework.Models;
 
 public record CommentNode
 {
