+++ CleannetCode_bot/Features/Homework/models/HomeworkServiceOptions.cs
@@ -1,4 +1,4 @@
-namespace CleannetCode_bot.Features.Homework;
+namespace CleannetCode_bot.Features.Homework.Models;
 
 public class HomeworkServiceOptions
 {
